opengl/requirements.txt
numpy
glfw
xvfbwrapper
PyOpenGL